Rework deployment
